MoPr

+++ Difficulty 2/coloring_socks.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 int main() {
-    int S, C, K, color;
+    long long S, C, K, color;
     cin >> S >> C >> K;
     vector<int> colors;
     for(int i = 0; i < S; i++) {
@@ -14,7 +14,7 @@ int main() {
 
     sort(colors.begin(), colors.end());
 
-    int i = 0, j = 1, machines = 1;
+    long long i = 0, j = 1, machines = 1;
     while(j < S) {
         if(colors.at(j) - colors.at(i) <= K && j - i <= C) j++;
         else {
